app/requirements.txt
fastapi 
uvicorn 
langchain 
langgraph 
sqlalchemy 
psycopg2-binary 
redis 
python-dotenv 
pydantic